+++ packages/journey-manager/tsconfig.json
@@ -3,7 +3,9 @@
   "extends": "./../../tsconfig.json",
   "compilerOptions": {
     "outDir": "./lib",
-    "rootDir": "./src"
+    "rootDir": "./src",
+    "jsx": "react-jsx",
+    "jsxImportSource": "preact",
   },
   "include": ["./src/**/*"]
 }
